+++ lib/widgets/buttons/exdock_copy_button.dart
